moved set_config() and update_version() back in the make_release.py module
